+ [ Memory optimization ]
*.c:
	- converted ints to short to reduce memory requirement

+++ 1/3/c2f.c
@@ -2,7 +2,7 @@
 
 main()
 {
-	int celsius;
+	short celsius;
 	printf("  °C\t    °F\n");
 	for (celsius = -100;celsius <= 100;celsius += 10)
 	{
